Finish ed10

+++ src/arranjo/validarArranjos.h
@@ -3,8 +3,8 @@
 
 #include<AEDI/arranjo.h>
 
-int validar_arranjosDiferentes(Arranjo* arranjo1, Arranjo* arranjo2);
-int validar_arranjo(Arranjo* arranjo);
+short validar_arranjosDiferentes(Arranjo* arranjo1, Arranjo* arranjo2);
+short validar_arranjo(Arranjo* arranjo);
 
 
 #endif
